+++ craft_parts/executor/part_handler.py
@@ -38,13 +38,14 @@ from craft_parts.parts import Part, get_parts_with_overlay, has_overlay_visibili
 from craft_parts.permissions import Permissions
 from craft_parts.plugins import Plugin
 from craft_parts.state_manager import MigrationState, StepState, states
+from craft_parts.state_manager.stage_state import StageState
 from craft_parts.steps import Step
 from craft_parts.utils import file_utils, os_utils
 
 from . import filesets, migration
 from .environment import generate_step_environment
 from .organize import organize_files
-from .step_handler import StepContents, StepHandler, Stream
+from .step_handler import StageContents, StepContents, StepHandler, Stream
 
 logger = logging.getLogger(__name__)
 
@@ -402,12 +403,15 @@ class PartHandler:
         """
         self._make_dirs()
 
-        contents = self._run_step(
-            step_info=step_info,
-            scriptlet_name="override-stage",
-            work_dir=self._part.stage_dir,
-            stdout=stdout,
-            stderr=stderr,
+        contents = cast(
+            StageContents,
+            self._run_step(
+                step_info=step_info,
+                scriptlet_name="override-stage",
+                work_dir=self._part.stage_dir,
+                stdout=stdout,
+                stderr=stderr,
+            ),
         )
 
         self._migrate_overlay_files_to_stage()
@@ -424,6 +428,8 @@ class PartHandler:
             files=contents.files,
             directories=contents.dirs,
             overlay_hash=overlay_hash.hex(),
+            backstage_files=contents.backstage_files,
+            backstage_directories=contents.backstage_dirs,
         )
 
     def _run_prime(
@@ -524,6 +530,8 @@ class PartHandler:
                 step=step_info.step,
                 work_dir=work_dir,
             )
+            if step_info.step == Step.STAGE:
+                return StageContents()
             return StepContents()
 
         return step_handler.run_builtin()
@@ -864,6 +872,8 @@ class PartHandler:
         for install_dir in self._part.part_install_dirs.values():
             _remove(install_dir)
 
+        _remove(self._part.part_export_dir)
+
     def _clean_stage(self) -> None:
         """Remove the current part's stage step files and state."""
         for (
@@ -871,6 +881,14 @@ class PartHandler:
             stage_dir,
         ) in self._part.stage_dirs.items():  # iterate over partitions
             self._clean_shared(Step.STAGE, partition=partition, shared_dir=stage_dir)
+
+        migration.clean_backstage(
+            part_name=self._part.name,
+            shared_dir=self._part.backstage_dir,
+            part_states=cast(
+                dict[str, StageState], _load_part_states(Step.STAGE, self._part_list)
+            ),
+        )
 
     def _clean_prime(self) -> None:
         """Remove the current part's prime step files and state."""
@@ -924,6 +942,7 @@ class PartHandler:
         dirs = [
             self._part.part_src_dir,
             self._part.part_build_dir,
+            self._part.part_export_dir,
             *self._part.part_install_dirs.values(),
             self._part.part_layer_dir,
             *self._part.part_layer_dirs.values(),