+++ tests/unit/features/overlay/test_parts.py
@@ -143,6 +143,7 @@ class TestPartData:
         assert p.part_layer_dir == new_dir / "parts/foo/layer"
         assert p.part_cache_dir == new_dir / "parts/foo/cache"
         assert p.overlay_dir == new_dir / "overlay"
+        assert p.backstage_dir == new_dir / "backstage"
         assert p.stage_dir == new_dir / "stage"
         assert p.prime_dir == new_dir / "prime"
 
@@ -165,6 +166,7 @@ class TestPartData:
         assert p.part_run_dir == new_dir / "foobar/parts/foo/run"
         assert p.part_layer_dir == new_dir / "foobar/parts/foo/layer"
         assert p.overlay_dir == new_dir / "foobar/overlay"
+        assert p.backstage_dir == new_dir / "foobar/backstage"
         assert p.stage_dir == new_dir / "foobar/stage"
         assert p.prime_dir == new_dir / "foobar/prime"