+++ docs/reference/changelog.rst
@@ -2,6 +2,13 @@
 Changelog
 *********
 
+2.10.0 (2025-MM-DD)
+-------------------
+
+- Add ``backstage`` and ``part/export`` directories for plugin use.
+
+For a complete list of commits, check out the `2.10.0`_ release on GitHub.
+
 2.9.0 (2025-04-02)
 ------------------
 
@@ -856,6 +863,7 @@ For a complete list of commits, check out the `2.0.0`_ release on GitHub.
 .. _craft-cli issue #172: https://github.com/canonical/craft-cli/issues/172
 .. _Poetry: https://python-poetry.org
 
+.. _2.10.0: https://github.com/canonical/craft-parts/releases/tag/2.10.0
 .. _2.4.2: https://github.com/canonical/craft-parts/releases/tag/2.4.2
 .. _2.4.0: https://github.com/canonical/craft-parts/releases/tag/2.4.0
 .. _2.3.1: https://github.com/canonical/craft-parts/releases/tag/2.3.1